renames article module file to articles
